pybricks/robotics/pb_type_matrix: use float instead of float_t
float_t is not always a 32-bit floating point type (it depends on the
value of FLT_EVAL_METHOD), however it is being used as if it is always
defined as float. Fix by using float type directly instead.

Also clean up some function names and parameters.

+++ pybricks/common/pb_type_lightgrid.c
@@ -68,7 +68,7 @@ static void common_LightGrid_image__extract(mp_obj_t image_in, size_t size, uint
     if (mp_obj_is_type(image_in, &pb_type_Matrix_type)) {
         for (size_t r = 0; r < size; r++) {
             for (size_t c = 0; c < size; c++) {
-                data[r * size + c] = pb_type_Matrix__get_scalar(image_in, r, c);
+                data[r * size + c] = (uint8_t)pb_type_Matrix__get_scalar(image_in, r, c);
             }
         }
         return;
